requirements.txt
django==1.7.4
djangorestframework>=3.0
mysql-python==1.2.5
Pillow==2.7.0
openpyxl==2.2.0
django-xmail-ritual>=0.0.11
django-dynsettings-ritual>=0.0.5
django-tinymce==1.5.3
sqlparse-0.1.15
pytz==2015.4
